+++ packages/reactabular-dnd/CHANGELOG.md
@@ -1,4 +1,4 @@
-7.0.0 / 2016-xx-xx
+7.0.0 / 2016-11-03
 ==================
 
   * Breaking - Rework `moveRows` interface. Instead of `moveRows(rows, { sourceRowId, targetRowId }) => rows` it's `moveRows({ sourceRowId, targetRowId }) => (rows) => <movedRows>` now.